Add files via upload
This is a project for a marketing department. This program will run leads against every contact and leads in the system. It will generate an output with a new column (_merge). The _merge column has two values “Both” and “left_only”. 
All the leads with “Left only” values are net new leads. In addition, it will also show the lead id or contact id if the record already exists as contact or a lead in SFDC.